move update absence from controller to the faculty service
