+++ src/server/faculty/faculty.service.ts
@@ -3,10 +3,19 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Faculty } from './faculty.entity';
 import { InstructorResponseDTO } from '../../common/dto/courses/InstructorResponse.dto';
 
+import { Absence } from 'server/absence/absence.entity';
+import { AbsenceResponseDTO } from 'common/dto/faculty/AbsenceResponse.dto';
+import { AbsenceRequestDTO } from 'common/dto/faculty/AbsenceRequest.dto';
+import { EntityNotFoundError } from 'typeorm/error/EntityNotFoundError';
+import { NotFoundException } from '@nestjs/common'
+
 export class FacultyService {
   @InjectRepository(Faculty)
   private readonly facultyRepository: Repository<Faculty>;
 
+  @InjectRepository(Absence)
+  private absenceRepository: Repository<Absence>;
+  
   /**
    * Retrieve all faculty members in the database with their associated area,
    * sorted by:
@@ -36,4 +45,30 @@ export class FacultyService {
       .orderBy('"displayName"')
       .getRawMany();
   }
+
+
+  public async updateFacultyAbsence( absenceInfo: AbsenceRequestDTO):
+  Promise<AbsenceResponseDTO> {
+    let existingAbsence: Absence;
+    try {
+      console.log (absenceInfo)
+      existingAbsence = await this.absenceRepository
+        .findOneOrFail({
+          where: {
+            id: absenceInfo.id,
+          },
+        });
+    } catch (e) {
+      if (e instanceof EntityNotFoundError) {
+        throw new NotFoundException('The entered Absence does not exist');
+      }
+      throw e;
+    }
+    const validAbsence = {
+      ...absenceInfo,
+      id: existingAbsence.id,
+    };
+    return this.absenceRepository.save(validAbsence);
+  }
+
 }
